+++ platform/darwin/src/NSExpression+MGLAdditions.mm
@@ -1,13 +1,12 @@
 #import "MGLFoundation_Private.h"
+#import "MGLGeometry_Private.h"
 #import "NSExpression+MGLPrivateAdditions.h"
 
 #import "MGLTypes.h"
 #if TARGET_OS_IPHONE
     #import "UIColor+MGLAdditions.h"
-    #define MGLEdgeInsets UIEdgeInsets
 #else
     #import "NSColor+MGLAdditions.h"
-    #define MGLEdgeInsets NSEdgeInsets
 #endif
 #import "NSPredicate+MGLAdditions.h"
 #import "NSValue+MGLStyleAttributeAdditions.h"